Projeto/backend/requirements.txt
flake8==6.1.0
black==23.11.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
gunicorn==21.2.0
django-debug-toolbar==4.2.0
django-cors-headers==4.3.1